// src/data/tokenMeta.generated.ts
export const Tokens = {
  BITCOIN: {
    BTC: { chain: "BTC", chainId: "BITCOIN", lightning: false, ticker: "BTC", decimals: 8, displayDecimals: 8, name: "Bitcoin (on-chain L1)", address: "", equals: (other: Token) => other.chainId === "BITCOIN" && other.ticker === "BTC", toString: (): "BTC" => "BTC" } satisfies BtcToken<false>,
    BTCLN: { chain: "BTC", chainId: "LIGHTNING", lightning: true, ticker: "BTC", decimals: 8, displayDecimals: 8, name: "Bitcoin (lightning L2)", address: "", equals: (other: Token) => other.chainId === "LIGHTNING" && other.ticker === "BTC", toString: (): "BTC-LN" => "BTC-LN" } satisfies BtcToken<true>,
  },
  SOLANA: {
    WBTC: { chain: "SC", chainId: "SOLANA", ticker: "WBTC", name: "Wrapped BTC (WBTC)", decimals: 8, displayDecimals: undefined, address: "3NZ9JMVBmGAqocybic2c7LQCJScmgsAZ6vQqTDzcqmJh", equals: (other: Token) => other.chainId === "SOLANA" && other.ticker === "WBTC" && other.address === "3NZ9JMVBmGAqocybic2c7LQCJScmgsAZ6vQqTDzcqmJh", toString: (): "SOLANA-WBTC" => "SOLANA-WBTC" } satisfies SCToken<"SOLANA">,
    USDC: { chain: "SC", chainId: "SOLANA", ticker: "USDC", name: "USD Circle", decimals: 6, displayDecimals: undefined, address: "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v", equals: (other: Token) => other.chainId === "SOLANA" && other.ticker === "USDC" && other.address === "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v", toString: (): "SOLANA-USDC" => "SOLANA-USDC" } satisfies SCToken<"SOLANA">,
    USDT: { chain: "SC", chainId: "SOLANA", ticker: "USDT", name: "Tether USD", decimals: 6, displayDecimals: undefined, address: "Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB", equals: (other: Token) => other.chainId === "SOLANA" && other.ticker === "USDT" && other.address === "Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB", toString: (): "SOLANA-USDT" => "SOLANA-USDT" } satisfies SCToken<"SOLANA">,
    SOL: { chain: "SC", chainId: "SOLANA", ticker: "SOL", name: "Solana", decimals: 9, displayDecimals: undefined, address: "So11111111111111111111111111111111111111112", equals: (other: Token) => other.chainId === "SOLANA" && other.ticker === "SOL" && other.address === "So11111111111111111111111111111111111111112", toString: (): "SOLANA-SOL" => "SOLANA-SOL" } satisfies SCToken<"SOLANA">,
    BONK: { chain: "SC", chainId: "SOLANA", ticker: "BONK", name: "Bonk", decimals: 5, displayDecimals: undefined, address: "DezXAZ8z7PnrnRJjz3wXBoRgixCa6xjnB7YaB1pPB263", equals: (other: Token) => other.chainId === "SOLANA" && other.ticker === "BONK" && other.address === "DezXAZ8z7PnrnRJjz3wXBoRgixCa6xjnB7YaB1pPB263", toString: (): "SOLANA-BONK" => "SOLANA-BONK" } satisfies SCToken<"SOLANA">,
  },
  STARKNET: {
    ETH: { chain: "SC", chainId: "STARKNET", ticker: "ETH", name: "Ether", decimals: 18, displayDecimals: 9, address: "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7", equals: (other: Token) => other.chainId === "STARKNET" && other.ticker === "ETH" && other.address === "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7", toString: (): "STARKNET-ETH" => "STARKNET-ETH" } satisfies SCToken<"STARKNET">,
    STRK: { chain: "SC", chainId: "STARKNET", ticker: "STRK", name: "Starknet", decimals: 18, displayDecimals: 9, address: "0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d", equals: (other: Token) => other.chainId === "STARKNET" && other.ticker === "STRK" && other.address === "0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d", toString: (): "STARKNET-STRK" => "STARKNET-STRK" } satisfies SCToken<"STARKNET">,
    WBTC: { chain: "SC", chainId: "STARKNET", ticker: "WBTC", name: "Wrapped BTC (WBTC)", decimals: 8, displayDecimals: undefined, address: "0x03fe2b97c1fd336e750087d68b9b867997fd64a2661ff3ca5a7c771641e8e7ac", equals: (other: Token) => other.chainId === "STARKNET" && other.ticker === "WBTC" && other.address === "0x03fe2b97c1fd336e750087d68b9b867997fd64a2661ff3ca5a7c771641e8e7ac", toString: (): "STARKNET-WBTC" => "STARKNET-WBTC" } satisfies SCToken<"STARKNET">,
    TBTC: { chain: "SC", chainId: "STARKNET", ticker: "TBTC", name: "Threshold BTC (tBTC)", decimals: 18, displayDecimals: 8, address: "0x04daa17763b286d1e59b97c283c0b8c949994c361e426a28f743c67bdfe9a32f", equals: (other: Token) => other.chainId === "STARKNET" && other.ticker === "TBTC" && other.address === "0x04daa17763b286d1e59b97c283c0b8c949994c361e426a28f743c67bdfe9a32f", toString: (): "STARKNET-TBTC" => "STARKNET-TBTC" } satisfies SCToken<"STARKNET">,
    USDC: { chain: "SC", chainId: "STARKNET", ticker: "USDC", name: "USD Circle", decimals: 6, displayDecimals: undefined, address: "0x033068f6539f8e6e6b131e6b2b814e6c34a5224bc66947c47dab9dfee93b35fb", equals: (other: Token) => other.chainId === "STARKNET" && other.ticker === "USDC" && other.address === "0x033068f6539f8e6e6b131e6b2b814e6c34a5224bc66947c47dab9dfee93b35fb", toString: (): "STARKNET-USDC" => "STARKNET-USDC" } satisfies SCToken<"STARKNET">,
    USDT: { chain: "SC", chainId: "STARKNET", ticker: "USDT", name: "Tether USD", decimals: 6, displayDecimals: undefined, address: "0x068f5c6a61780768455de69077e07e89787839bf8166decfbf92b645209c0fb8", equals: (other: Token) => other.chainId === "STARKNET" && other.ticker === "USDT" && other.address === "0x068f5c6a61780768455de69077e07e89787839bf8166decfbf92b645209c0fb8", toString: (): "STARKNET-USDT" => "STARKNET-USDT" } satisfies SCToken<"STARKNET">,
    strkBTC: { chain: "SC", chainId: "STARKNET", ticker: "strkBTC", name: "strkBTC", decimals: 8, displayDecimals: undefined, address: "0x0787150e306e6eae6e3f79dea881770e8bbff2c1b8eb490f969669ee945b3135", equals: (other: Token) => other.chainId === "STARKNET" && other.ticker === "strkBTC" && other.address === "0x0787150e306e6eae6e3f79dea881770e8bbff2c1b8eb490f969669ee945b3135", toString: (): "STARKNET-strkBTC" => "STARKNET-strkBTC" } satisfies SCToken<"STARKNET">,
    _TESTNET_WBTC_VESU: { chain: "SC", chainId: "STARKNET", ticker: "_TESTNET_WBTC_VESU", name: "Wrapped BTC (WBTC)", decimals: 8, displayDecimals: undefined, address: "0x04861ba938aed21f2cd7740acd3765ac4d2974783a3218367233de0153490cb6", equals: (other: Token) => other.chainId === "STARKNET" && other.ticker === "_TESTNET_WBTC_VESU" && other.address === "0x04861ba938aed21f2cd7740acd3765ac4d2974783a3218367233de0153490cb6", toString: (): "STARKNET-_TESTNET_WBTC_VESU" => "STARKNET-_TESTNET_WBTC_VESU" } satisfies SCToken<"STARKNET">,
    _TESTNET_strkBTC: { chain: "SC", chainId: "STARKNET", ticker: "_TESTNET_strkBTC", name: "strkBTC (dev)", decimals: 8, displayDecimals: undefined, address: "0x018a6bbc4b0c05135af4a1ff8251687d689b64609bbd79b7cb044f410e0a8ab2", equals: (other: Token) => other.chainId === "STARKNET" && other.ticker === "_TESTNET_strkBTC" && other.address === "0x018a6bbc4b0c05135af4a1ff8251687d689b64609bbd79b7cb044f410e0a8ab2", toString: (): "STARKNET-_TESTNET_strkBTC" => "STARKNET-_TESTNET_strkBTC" } satisfies SCToken<"STARKNET">,
  },
// ...
} satisfies Record<string, Record<string, Token>>;